requirements.txt
torch
trochvision
numpy
pillow
Scipy
flask
werkzeug
